moved venv functions to utils
